+++ packages/yew/src/lib.rs
@@ -263,10 +263,9 @@ pub mod macros {
     pub use crate::props;
 }
 
-mod app_handle;
 pub mod callback;
 pub mod context;
-pub mod dom_bundle;
+mod dom_bundle;
 pub mod functional;
 pub mod html;
 mod io_coop;
@@ -275,12 +274,15 @@ mod sealed;
 #[cfg(feature = "ssr")]
 mod server_renderer;
 pub mod suspense;
-#[cfg(test)]
-pub mod tests;
 pub mod utils;
 pub mod virtual_dom;
 #[cfg(feature = "ssr")]
 pub use server_renderer::*;
+
+#[cfg(test)]
+pub mod tests {
+    pub use crate::dom_bundle::layout_tests;
+}
 
 /// The module that contains all events available in the framework.
 pub mod events {
@@ -295,7 +297,7 @@ pub mod events {
     };
 }
 
-pub use crate::app_handle::AppHandle;
+pub use crate::dom_bundle::AppHandle;
 use web_sys::Element;
 
 use crate::html::BaseComponent;
@@ -375,9 +377,9 @@ where
 /// use yew::prelude::*;
 /// ```
 pub mod prelude {
-    pub use crate::app_handle::AppHandle;
     pub use crate::callback::Callback;
     pub use crate::context::ContextProvider;
+    pub use crate::dom_bundle::AppHandle;
     pub use crate::events::*;
     pub use crate::html::{
         create_portal, BaseComponent, Children, ChildrenWithProps, Classes, Component, Context,
